Convert URL operations
Delete legacy URL module
